# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SeuDrink API",
        "description": "Documentação da API usando Swagger",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/api/bebidas": {
            "get": {
                "tags": [
                    "Bebidas"
                ],
                "summary": "Lista todas as bebidas filtradas",
                "operationId": "37ccffc612a2280f9619233d3ac8795c",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Nome da bebida",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de bebidas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Bebida"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/bebida/novo": {
            "post": {
                "tags": [
                    "Bebidas"
                ],
                "summary": "Cria uma nova bebida",
                "operationId": "d33eff028d0f1b14627c24a507c365ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Vodka"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bebida criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bebida"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/bebida/{id}": {
            "put": {
                "tags": [
                    "Bebidas"
                ],
                "summary": "Atualiza uma bebida existente",
                "operationId": "1972c1995d8a87003729b71178f3d8e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da bebida",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Cerveja"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bebida atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bebida"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bebida não encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Bebida": {
                "title": "Bebida",
                "description": "Modelo de uma Bebida",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Vodka"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}